backend/requirements.txt
# Django and DRF
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0

# Redis
redis>=4.5,<5.0
django-redis>=5.4,<6.0

# ASGI/WebSockets
channels>=4.0,<5.0
daphne>=4.0,<5.0

# HTTP Client
requests>=2.31,<3.0

# Environment variables
python-dotenv>=1.0,<2.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
responses>=0.23,<1.0

# Production
gunicorn>=21.0,<22.0
